fixed according to review

+++ src/main/java/ru/yandex/practicum/filmorate/services/LikeService.java
@@ -2,9 +2,8 @@ package ru.yandex.practicum.filmorate.services;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.yandex.practicum.filmorate.dao.FilmDao;
 import ru.yandex.practicum.filmorate.dao.LikeDao;
-import ru.yandex.practicum.filmorate.dao.UserDao;
+import ru.yandex.practicum.filmorate.exception.ObjectNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import javax.validation.ValidationException;
@@ -15,16 +14,15 @@ import java.util.List;
 public class LikeService {
 
     private final LikeDao likeDao;
-    private final UserDao userDao;
-    private final FilmDao filmDao;
 
     public void add(long filmId, long userId) {
         likeDao.add(userId, filmId);
     }
 
     public void delete(long filmId, long userId) {
-        filmDao.get(filmId);
-        userDao.get(userId);
+        if (userId <= 0 || filmId <= 0) {
+            throw new ObjectNotFoundException("Объект не найден");
+        }
         likeDao.delete(filmId, userId);
     }
